requirements.txt
fpdf
pandas
python-docx
markdown2
pdfkit
python-pptx